fix(land): thirteen findings, and three of them were gates deciding nothing
Review of #848, worked as one batch. The three classes worth naming, because
each recurred rather than appeared once:

A SHARED THING CHANGED TO FIX ONE CALLER'S SYMPTOM.
- `exec.rs` merged stderr at the shared spawn so a landing gate could report
  its cause. `review.rs` and `recorder.rs` parse that stream, so both broke.
  It is `Diagnostics { Drop, Keep }` per call site now: `piped` drops,
  `piped_argv` keeps.
- `credential()` filtered AFTER the fold, so an exported-but-EMPTY `GH_TOKEN`
  suppressed the `GITHUB_TOKEN` fallback rather than falling through to it.

A MECHANISM DECLARED AND NEVER REACHED.
- `fast_forward::grade` mapped `"skipped"` onto `Answer::Accepted`, which
  reached `Progress::Landed` and deleted the remote branch under an OPEN PR.
- `landed_for_real` asked the forge with `state=open`. A merged PR is closed,
  so `Merged::Yes` was unreachable and the branch could never land. It reads
  the PR's own `merged` boolean through `pull_request_in_any_state` now, with
  `Lookup::{Found, None, Unreadable}` keeping could-not-look apart from no-PR.
- `task.rs` `continue`d past an unreadable registry entry and reported
  "nothing registered" at exit 0. `Reading::Unreadable(path)` instead.
- `landed::decide` asked the landed arm first, so an explicit human
  `DO-NOT-CLOSE` was answered with "advance it" — the inversion of the arm
  this module's own doc calls load-bearing. The decline is asked first, and
  the two arms are an `else if` chain so the exclusion is structural.
- `fixture-forks.rego` had no could-not-look clause; a base rev that did not
  resolve read as a clean tree.

COPYING A NEARBY IDIOM INSTEAD OF READING WHAT IT MEANS.
- `hook::qualify` was handed raw tokens rather than `program_token`-normalised
  ones, so `--in-place)` defeated `requires_flag`.
- The fast-forward answer poll grew a SECOND `std::thread::sleep` with its own
  `disallowed_methods` escape, ~forty lines from `land.rs`'s comment saying
  there is one sleep in this crate and a second arm may not grow a timer of
  its own. It calls `pr_watch::pause_until` now, and `spawn-widening` is what
  caught it rather than any reader.
- `spawn-placement-entry` justified its narrowness by asserting the placement
  table is "one per line". `spawn-adapters.rego:151-153` packs four to six
  names per line, so appending to an existing line evaded clause B entirely.

The rest, each its own arithmetic or reading defect:
- `rest.rs` sent no `User-Agent`, which GitHub 403s. `gh` sent one for free,
  so the retirement inherited nothing and every API read failed.
- `backoff_of` read a failed clock (`now_unix() == 0`) as an instant, passed
  the `reset > now` filter and yielded the raw epoch — a 57-year sleep. Now
  guarded, and capped by `MAX_BACKOFF`.
- `lease.rs` `delta_size`/`pack_header` shifted by an unbounded amount taken
  from remote bytes: a debug panic, a release mis-decode. `checked_shl`.
- `lease::terms` restores `beat = ttl / BEATS_PER_TTL` where a configured beat
  meets or exceeds the TTL, so a heartbeat cannot be scheduled after expiry.
- `lease::health` computes `expires - now` with `checked_sub` and reports
  `Wedged` for a body whose arithmetic will not close.
- `lease::fetch` qualifies a bare ref as `refs/heads/{ref}`.
- `fetch::exchange` strips `authorization` across a host change on redirect.
- `pr_watch`/`main_watch` `Poll::absorb` read `answer.is_reading()` rather
  than `status != 304`, so an error page no longer updates the ETag.

`batten.toml` is a protected path; the pattern widening carries its admission.

Refs: CLOUD-1148

Admits: 3faa3c73a28dd326c47ea79e83467529d728dea3fd9fc5aacb829ad88da8c541
Admits-rule: protected-mutation
Admits-verdict: path write refused
Admits-subject: batten.toml
Admits-anchor: call:abf6abc1f0c54472335e67291d204d7d6745e6a8
Admits-epoch: acdfb7d1bc053d9f6b347d1414059354e9ebeee4386a0d371f939ac41b748a90
Admits-author: alec@wenzowski.com
Admits-prev: 97dbde60d0a0de2085a42a81286d7a5a333dec0b44ac20e99300a6e52f3550fa
Admits-answer-lost: Nothing. The row still matches only lines made entirely of quoted lower-snake words with commas, which is the set's shape and not prose or a fixture — the concern the original comment names. It matches strictly more of the real table than before.
Admits-answer-precondition: `spawn-placement-entry` is `^\s*"[a-z_]+",\s*$` — one quoted name ALONE on its line — and its comment justifies that narrowness by asserting the placement table is "a Rego set of bare module names, one per line". That is false of the file it reads: `policy/spawn-adapters.rego:151-153` packs four to six names per line, which is the file's dominant style. So widening the table by appending to an existing line produces an added line the pattern does not match, `spawn-widening`'s clause B never fires, and the gate is evaded by a formatting choice rather than by an argument.
Admits-answer-rejected-route: Reformatting `spawn-adapters.rego` to one name per line so the existing pattern holds. That makes the gate depend on a formatting convention no gate enforces, so the next author who packs a line silently reopens the hole — trading a pattern defect for a latent one.

+++ crates/batten/src/fast_forward.rs
@@ -388,7 +388,13 @@ pub fn answer_request(ask: &Ask, since: &str, page: u32) -> String {
 /// **The loop is the caller's, never this module's**, for the reason
 /// [`crate::main_watch`]'s header gives about its own: a module holding an
 /// unbounded loop becomes a second authority over when to stop asking.
-pub const ANSWER_POLL_SECONDS: u64 = 2;
+///
+/// Spelled `f64` because the one sleep in this crate is
+/// [`crate::pr_watch::pause_until`] and that is what it takes. A second timer
+/// here would need a second `disallowed_methods` escape, which is the widening
+/// `spawn-widening` refuses and which `land.rs`'s stale arm already declines for
+/// the same reason.
+pub const ANSWER_POLL_SECONDS: f64 = 2.0;
 
 #[must_use]
 pub fn answer(ask: &Ask, since: &str, comment: &str) -> Answer {
